refactor(tests): dedupe ios defaults and setup-code helpers
(cherry picked from commit 500883775baea3698531fc9e4a25fe2e4ee1d81b)
